backend/requirements.txt
fastapi
pydantic
python-dotenv
python-multipart
httpx

